+++ src/image.rs
@@ -221,6 +221,7 @@ mod tests {
                 code_block_width: 4,
                 code_block_height: 4,
                 code_block_style: 0,
+                multiple_component_transform: false,
                 transform: Transform::Reversible53,
                 precinct_sizes: vec![],
             },